+++ NEWS.md
@@ -1,6 +1,14 @@
+# pastclim 1.2.3
+* Added lai to Krapp2021 (the variable is now also present in the original OSF
+  repository for that dataset).
+* Change column names in data.frame returned by location_series to match
+  location_slice
+* Allow for interpolation of nearest neighbours in location_series, and allow
+  for a buffer on estimates returned by the location_* functions.
+
 # pastclim 1.2.2
 * Update of Krapp2021 files to make them compatible with how terra now handles
-  time. Users will have to redownload datasets. Old files can be removed with
+  time. Users will have to re-download datasets. Old files can be removed with
   'clean_data_path()'
 
 # pastclim 1.2.1